refactor: move HashGenerator and Authentication classes to Domain\Identity\Service namespace
