+++ tests/Integration/Domain/Identity/Service/HashGeneratorTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PhpList\Core\Tests\Integration\Security;
+namespace PhpList\Core\Tests\Integration\Domain\Identity\Service;
 
 use Doctrine\ORM\Tools\SchemaTool;
-use PhpList\Core\Security\HashGenerator;
+use PhpList\Core\Domain\Identity\Service\HashGenerator;
 use PhpList\Core\TestingSupport\Traits\DatabaseTestTrait;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
