+++ tests/Unit/Domain/Identity/Service/AdministratorManagerTest.php
@@ -9,7 +9,7 @@ use PhpList\Core\Domain\Identity\Model\Administrator;
 use PhpList\Core\Domain\Identity\Model\Dto\CreateAdministratorDto;
 use PhpList\Core\Domain\Identity\Model\Dto\UpdateAdministratorDto;
 use PhpList\Core\Domain\Identity\Service\Manager\AdministratorManager;
-use PhpList\Core\Security\HashGenerator;
+use PhpList\Core\Domain\Identity\Service\HashGenerator;
 use PHPUnit\Framework\TestCase;
 
 class AdministratorManagerTest extends TestCase
